backend/requirements.txt
Django>=5.0,<5.1
django-ninja>=1.1.0
django-cors-headers>=4.3.0
